charts: Add support for PVC 'spec.selector'
When pre-provisioning PersistentVolumes for a Zenko deployment, one may
not want to (and should not) use one StorageClass per StatefulSet, and
instead will likely want to rely on some form of selector to match PVs
with the PVCs generated by the StatefulSet.
To support this, we add this option to each chart supporting persistent
storage, either under `.persistentVolume.selector` or
`.persistence.selector`, depending on the individual chart.

Fixes: ZENKO-2424